+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2023-06-8
+
+### Changed
+
+- Updated pubspec.yaml to have a more accurate description of the library.
+- Fixed name of the class from `Snoflaker` to `Snowflaker`.
+
+### Added
+
+- Created example
+
 ## [1.0.0] - 2023-06-08
 
 ### Added